# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/users/":{"get":{"tags":["users"],"summary":"Read Users","description":"Read users.\n\nDoesn't require authentication.","operationId":"read_users_users__get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Users Users  Get","type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["users"],"summary":"Create User","description":"Create a new user. Ony accessible by the superuser.","operationId":"create_user_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["users"],"summary":"Read User Me","description":"Read the user data for the currently authenticated user.","operationId":"read_user_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/{user_id}":{"get":{"tags":["users"],"summary":"Read User","description":"Read the data for a specific user.\n\nDoesn't require authentication.","operationId":"read_user_users__user_id__get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/auctions/":{"get":{"tags":["auctions"],"summary":"Read User Auctions","description":"Read all the auctions of a specific user. Doesn't need authentication.","operationId":"read_user_auctions_users__user_id__auctions__get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read User Auctions Users  User Id  Auctions  Get","type":"array","items":{"$ref":"#/components/schemas/Auction"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["auctions"],"summary":"Create Auction For User","description":"Create an auction\n\nRequires authentication and the auction will be assigned to the current user.","operationId":"create_auction_for_user_users__user_id__auctions__post","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuctionCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Auction"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auctions/":{"get":{"tags":["auctions"],"summary":"Read Auctions","description":"Read all the auctions. Doesn't need authentication.","operationId":"read_auctions_auctions__get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Auctions Auctions  Get","type":"array","items":{"$ref":"#/components/schemas/Auction"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auctions/{auction_id}/":{"get":{"tags":["auctions"],"summary":"Read Auction","description":"Read a auction. Doesn't nedd authentication.","operationId":"read_auction_auctions__auction_id___get","parameters":[{"required":true,"schema":{"title":"Auction Id","type":"integer"},"name":"auction_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Auction"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auctions/{auction_id}/items/":{"get":{"tags":["items"],"summary":"Read Auction Items","description":"Get the items of an auction.","operationId":"read_auction_items_auctions__auction_id__items__get","parameters":[{"required":true,"schema":{"title":"Auction Id","type":"integer"},"name":"auction_id","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Auction Items Auctions  Auction Id  Items  Get","type":"array","items":{"$ref":"#/components/schemas/Item"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["items"],"summary":"Create Item For Auction","description":"Create an item for an auction\n\nRequires authentication and the item will be assigned to the auction. The current user must the owner of the auction.","operationId":"create_item_for_auction_auctions__auction_id__items__post","parameters":[{"required":true,"schema":{"title":"Auction Id","type":"integer"},"name":"auction_id","in":"path"},{"required":true,"schema":{"title":"Owner Id","type":"integer"},"name":"owner_id","in":"query"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/items/":{"get":{"tags":["items"],"summary":"Read All Items","description":"Read all items.","operationId":"read_all_items_items__get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read All Items Items  Get","type":"array","items":{"$ref":"#/components/schemas/Item"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/items/{item_id}/":{"get":{"tags":["items"],"summary":"Read Item","description":"Read an items.","operationId":"read_item_items__item_id___get","parameters":[{"required":true,"schema":{"title":"Item Id","type":"integer"},"name":"item_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/items/":{"get":{"tags":["items"],"summary":"Read User Items","description":"Read an user's items.","operationId":"read_user_items_users__user_id__items__get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read User Items Users  User Id  Items  Get","type":"array","items":{"$ref":"#/components/schemas/Item"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/items/{item_id}/bids":{"post":{"tags":["bids"],"summary":"Create Bid For Item","description":"Create an Bid for a specific item.\n\nOnly allowed to an authenticated user. He will be it owner\n\nAlso the item must exist.","operationId":"create_bid_for_item_items__item_id__bids_post","parameters":[{"required":true,"schema":{"title":"Item Id","type":"integer"},"name":"item_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BidCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Bid"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/items/{item_id}/bids/":{"get":{"tags":["bids"],"summary":"Read Bids For Item","description":"Read all the bids for a specific item.\nDoes'nt require authentication.\nThe item must exist.","operationId":"read_bids_for_item_items__item_id__bids__get","parameters":[{"required":true,"schema":{"title":"Item Id","type":"integer"},"name":"item_id","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Bids For Item Items  Item Id  Bids  Get","type":"array","items":{"$ref":"#/components/schemas/Bid"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/bids/":{"get":{"tags":["bids"],"summary":"Read All Bids","description":"Read all bids\nDoes'nt require authentication.","operationId":"read_all_bids_bids__get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read All Bids Bids  Get","type":"array","items":{"$ref":"#/components/schemas/Bid"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/bids/{bid_id}/":{"get":{"tags":["bids"],"summary":"Read Bid","description":"Read a bid. Doesn't nedd authentication.","operationId":"read_bid_bids__bid_id___get","parameters":[{"required":true,"schema":{"title":"Bid Id","type":"integer"},"name":"bid_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Bid"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me/bids/":{"get":{"tags":["bids"],"summary":"Read My Bids","description":"Read all my bids.","operationId":"read_my_bids_users_me_bids__get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read My Bids Users Me Bids  Get","type":"array","items":{"$ref":"#/components/schemas/Bid"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/{user_id}/bids/":{"get":{"tags":["bids"],"summary":"Read User Bids","description":"Read all the bids of a specific user.","operationId":"read_user_bids_users__user_id__bids__get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read User Bids Users  User Id  Bids  Get","type":"array","items":{"$ref":"#/components/schemas/Bid"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me/bids/{auction_id}":{"get":{"tags":["bids"],"summary":"Read My Bids For Auction","description":"Read all my bids for an auction","operationId":"read_my_bids_for_auction_users_me_bids__auction_id__get","parameters":[{"required":true,"schema":{"title":"Auction Id","type":"integer"},"name":"auction_id","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read My Bids For Auction Users Me Bids  Auction Id  Get","type":"array","items":{"$ref":"#/components/schemas/Bid"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/{user_id}/bids/item":{"get":{"tags":["bids"],"summary":"Read User Bids For Item","description":"Read all the bids of a specific an user for an item","operationId":"read_user_bids_for_item_users__user_id__bids_item_get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"},{"required":true,"schema":{"title":"Item Id","type":"integer"},"name":"item_id","in":"query"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read User Bids For Item Users  User Id  Bids Item Get","type":"array","items":{"$ref":"#/components/schemas/Bid"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login/access-token":{"post":{"summary":"Login Access Token","description":"OAuth2 compatible token login, get an access token for future requests.","operationId":"login_access_token_login_access_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_access_token_login_access_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Auction":{"title":"Auction","required":["title","id","owner_id","creation_date"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"start_date":{"title":"Start Date","type":"string","format":"date-time"},"is_active":{"title":"Is Active","type":"boolean","default":false},"id":{"title":"Id","type":"integer"},"owner_id":{"title":"Owner Id","type":"integer"},"creation_date":{"title":"Creation Date","type":"string","format":"date-time"},"items":{"title":"Items","type":"array","items":{"$ref":"#/components/schemas/Item"},"default":[]}}},"AuctionCreate":{"title":"AuctionCreate","required":["title"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"start_date":{"title":"Start Date","type":"string","format":"date-time"},"is_active":{"title":"Is Active","type":"boolean","default":false}}},"Bid":{"title":"Bid","required":["price","id","bidder_id","item_id","bid_date"],"type":"object","properties":{"price":{"title":"Price","type":"integer"},"id":{"title":"Id","type":"integer"},"bidder_id":{"title":"Bidder Id","type":"integer"},"item_id":{"title":"Item Id","type":"integer"},"bid_date":{"title":"Bid Date","type":"string","format":"date-time"}}},"BidCreate":{"title":"BidCreate","required":["price"],"type":"object","properties":{"price":{"title":"Price","type":"integer"}}},"Body_login_access_token_login_access_token_post":{"title":"Body_login_access_token_login_access_token_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Item":{"title":"Item","required":["title","description","start_price","id","owner_id","auction_id"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"start_price":{"title":"Start Price","type":"integer"},"picture_url":{"title":"Picture Url","type":"string"},"id":{"title":"Id","type":"integer"},"owner_id":{"title":"Owner Id","type":"integer"},"auction_id":{"title":"Auction Id","type":"integer"},"bids":{"title":"Bids","type":"array","items":{"$ref":"#/components/schemas/Bid"},"default":[]}}},"ItemCreate":{"title":"ItemCreate","required":["title","description","start_price"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"start_price":{"title":"Start Price","type":"integer"},"picture_url":{"title":"Picture Url","type":"string"}}},"Token":{"title":"Token","required":["access_token","token_type"],"type":"object","properties":{"access_token":{"title":"Access Token","type":"string"},"token_type":{"title":"Token Type","type":"string"}}},"User":{"title":"User","required":["first_name","last_name","email","id","is_active"],"type":"object","properties":{"first_name":{"title":"First Name","type":"string"},"last_name":{"title":"Last Name","type":"string"},"email":{"title":"Email","type":"string"},"id":{"title":"Id","type":"integer"},"is_active":{"title":"Is Active","type":"boolean"},"auctions":{"title":"Auctions","type":"array","items":{"$ref":"#/components/schemas/Auction"},"default":[]}}},"UserCreate":{"title":"UserCreate","required":["first_name","last_name","email","password"],"type":"object","properties":{"first_name":{"title":"First Name","type":"string"},"last_name":{"title":"Last Name","type":"string"},"email":{"title":"Email","type":"string"},"password":{"title":"Password","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/login/access-token"}}}}}}